+++ src/plugins/assert/otpl.assert.js
@@ -4,10 +4,8 @@ utils.has = function ( data, key, type ) {
 
 	if ( !data ) return false;
 
-	if ( !type && data[ key ] != undefined ) return true;
-	if ( data[ key ] != undefined && utils.type( data[ key ], type ) ) return true;
-
-	return false;
+	return ( !type && data[ key ] != undefined )
+	|| ( data[ key ] != undefined && utils.type( data[ key ], type ) );
 };
 
 utils.type = function ( value, type ) {
@@ -15,11 +13,11 @@ utils.type = function ( value, type ) {
 
 	switch ( type ) {
 		case 'string':
-			ans = ( typeof value === 'string') ? true : false;
+			ans = ( typeof value === 'string');
 			break;
 		case 'list':
 			var str = Object.prototype.toString.call( value );
-			ans     = ( str === '[object Array]' || str === '[object Object]' ) ? true : false;
+			ans     = ( str === '[object Array]' || str === '[object Object]' );
 			break;
 		case 'numeric':
 			ans = !isNaN( value );